+++ crates/fuzz/src/lib.rs
@@ -1,9 +1,7 @@
 pub mod accounts_storage;
 pub mod error;
 pub mod fuzz_client_impl;
-pub mod fuzz_data;
 pub mod fuzz_stats;
-pub mod macros;
 pub mod traits;
 
 pub mod trident_accounts;
@@ -36,30 +34,31 @@ pub mod fuzzing {
     pub use arbitrary::Arbitrary;
     pub use honggfuzz::fuzz as fuzz_honggfuzz;
 
-    /// trident derive
-    pub use trident_derive_accounts::TridentAccounts;
-    pub use trident_derive_fuzz_test_executor::FuzzTestExecutor;
-    pub use trident_derive_instruction::TridentInstruction;
-    pub use trident_derive_remaining_accounts::TridentRemainingAccounts;
-    pub use trident_derive_transaction::TridentTransaction;
-
     /// trident traits
     pub use super::traits::AccountsMethods;
     pub use super::traits::FuzzClient;
-    pub use super::traits::FuzzTestExecutor;
     pub use super::traits::InstructionMethods;
     pub use super::traits::InstructionSetters;
     pub use super::traits::RemainingAccountsMethods;
     pub use super::traits::TransactionCustomMethods;
     pub use super::traits::TransactionGetters;
     pub use super::traits::TransactionMethods;
+    pub use super::traits::TransactionSelector;
     pub use super::traits::TransactionSetters;
+    /// trident derive
+    pub use trident_derive_accounts::TridentAccounts;
+    pub use trident_derive_flow_executor::flow;
+    pub use trident_derive_flow_executor::flow_executor;
+    pub use trident_derive_flow_executor::flow_ignore;
+    pub use trident_derive_flow_executor::init;
+
+    pub use trident_derive_fuzz_test_executor::FuzzTestExecutor;
+    pub use trident_derive_instruction::TridentInstruction;
+    pub use trident_derive_remaining_accounts::TridentRemainingAccounts;
+    pub use trident_derive_transaction::TridentTransaction;
+    pub use trident_derive_transaction_selector::TransactionSelector;
     /// trident svm
     pub use trident_svm::processor;
-
-    /// trident macros
-    pub use super::fuzz_trident;
-    pub use super::sequence;
 
     /// accounts storages
     pub use super::accounts_storage::account_storage::AccountsStorage;
@@ -69,8 +68,6 @@ pub mod fuzzing {
     pub use trident_config::TridentConfig;
 
     pub use super::error::*;
-    pub use super::fuzz_data::build_ix_fuzz_data;
-    pub use super::fuzz_data::*;
     pub use super::fuzz_stats::FuzzingStatistics;
 
     pub use std::cell::RefCell;
@@ -81,11 +78,12 @@ pub mod fuzzing {
     /// types
     pub use crate::types::AccountId;
     pub use crate::types::FuzzerData;
-    pub use crate::types::SequenceResult;
 
     /// trident accounts
     pub use crate::trident_accounts::TridentAccount;
     pub use crate::trident_pubkey::TridentPubkey;
 
     pub use borsh::{BorshDeserialize, BorshSerialize};
+
+    pub use arbitrary::Unstructured;
 }