+++ crates/syn/src/codegen/trident_fuzz_test_executor.rs
@@ -1,32 +1,48 @@
 use proc_macro2::TokenStream;
 use quote::{quote, ToTokens};
 
-use crate::types::trident_fuzz_test_executor::TridentFuzzTestExecutorEnum;
+use crate::types::trident_fuzz_test_executor::TridentFuzzTestExecutor;
 
-impl ToTokens for TridentFuzzTestExecutorEnum {
+impl ToTokens for TridentFuzzTestExecutor {
     fn to_tokens(&self, tokens: &mut TokenStream) {
         let name = &self.ident;
-        let variants = &self.variants;
-
-        let process_transaction_match_arms = variants.iter().map(|variant| {
-            let variant_name = &variant.ident;
-            quote! {
-                #name::#variant_name(tx) => {
-                    tx.process_transaction(client, config, &mut fuzz_accounts.borrow_mut())
-                }
-            }
-        });
+        let (impl_generics, ty_generics, _) = self.generics.split_for_impl();
+        let client_type = &self.client_type;
 
         let expanded = quote! {
-            impl FuzzTestExecutor<FuzzAccounts> for #name {
-                fn process_transaction(
-                    &mut self,
-                    client: &mut TridentSVM,
-                    config: &TridentConfig,
-                    fuzz_accounts: &RefCell<FuzzAccounts>,
-                ) -> Result<(), FuzzingError> {
-                    match self {
-                        #(#process_transaction_match_arms)*
+            impl #impl_generics #name #ty_generics {
+                fn new(client: #client_type, config: TridentConfig) -> Self {
+                    Self { client, config }
+                }
+
+                fn fuzz(&mut self) {
+                    if cfg!(honggfuzz) {
+                        loop {
+                            fuzz_honggfuzz(|fuzzer_data| {
+                                let mut buf = Unstructured::new(fuzzer_data);
+                                let _ = self.execute_flows(&mut buf);
+                                self.client.clear_accounts();
+                            });
+                        }
+                    } else if cfg!(afl) {
+                        fuzz_afl(true, |fuzzer_data| {
+                            let mut buf = Unstructured::new(fuzzer_data);
+                            let _ = self.execute_flows(&mut buf);
+                            self.client.clear_accounts();
+                        });
+                    } else if cfg!(honggfuzz_debug) {
+                        let mut crash_file = String::new();
+                        std::io::stdin()
+                            .read_line(&mut crash_file)
+                            .expect("Failed to read crash file path from stdin");
+                        let crash_file = crash_file.trim();
+
+                        let fuzzer_data = std::fs::read(crash_file).expect("Failed to read crash file");
+                        let mut buf = Unstructured::new(&fuzzer_data);
+                        let _ = self.execute_flows(&mut buf);
+                        self.client.clear_accounts();
+                    } else {
+                        panic!("Select Honggfuzz or AFL for fuzzing!!!")
                     }
                 }
             }