+++ crates/template/src/template_getters.rs
@@ -71,7 +71,7 @@ impl Template {
             /// You can create your own transactions by adding new variants to the enum.
             ///
             /// Docs: https://ackee.xyz/trident/docs/latest/trident-api-macro/trident-types/fuzz-transactions/
-            #[derive(Arbitrary, FuzzTestExecutor)]
+            #[derive(Arbitrary, TransactionSelector)]
             pub enum FuzzTransactions {
                 #(#transaction_variants),*
             }